requirements.txt
django==3.1.5
django-debug-toolbar==3.2
dj-database-url==0.5.0
Pillow==8.1.0
environs==9.3.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
geopy==2.1.0
requests~=2.25.1
